fix: press Escape to exit ImageBuildLog
